agrosathi-model-main/requirements.txt
gunicorn
flask
numpy
pillow
tensorflow==2.13.0
requests
flask-cors